requirements.txt
numpy~=1.21.2
pandas~=1.3.4
scipy~=1.7.3
statsmodels~=0.13.1
matplotlib~=3.4.3
seaborn~=0.11.2
tqdm~=4.62.3
biopython~=1.79
ete3~=3.1.2
click~=8.0.3
networkx~=2.6.3